+++ src/SchemaKeeper/Outside/DeployedFunctions.php
@@ -5,7 +5,7 @@
  * For the full copyright and license information, please view the LICENSE file that was distributed with this source code.
  */
 
-namespace SchemaKeeper\Worker;
+namespace SchemaKeeper\Outside;
 
 class DeployedFunctions
 {
